+++ library/PSFramework/Data/Psd1Converter.cs
@@ -3,8 +3,10 @@ using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
 using System.Management.Automation;
+using System.Management.Automation.Language;
 using System.Text;
 using System.Threading.Tasks;
+using static System.Net.Mime.MediaTypeNames;
 
 namespace PSFramework.Data
 {
@@ -34,6 +36,11 @@ namespace PSFramework.Data
         public Hashtable Config = new Hashtable(StringComparer.OrdinalIgnoreCase);
 
         /// <summary>
+        /// Enable verbose messages.
+        /// </summary>
+        public bool EnableVerbose;
+
+        /// <summary>
         /// Convert some object to PSD1
         /// </summary>
         /// <param name="Value">The object to convert to PSD1</param>
@@ -44,11 +51,26 @@ namespace PSFramework.Data
         }
 
         /// <summary>
+        /// Convert some object to PSD1
+        /// </summary>
+        /// <param name="Value">The object to convert to PSD1</param>
+        /// <param name="Depth">How deeply indented do we start?</param>
+        /// <param name="Parents">What parent objects came before us? Helps prevent infinite recursion.</param>
+        /// <returns>A PSD1-string</returns>
+        public string Convert(object Value, object[] Parents, int Depth)
+        {
+            return DataHost.Convert(Value, Parents, Depth, this);
+        }
+
+        /// <summary>
         /// Send a message (maybe)
         /// </summary>
         /// <param name="Message"></param>
         public void WriteVerbose(string Message)
         {
+            if (!EnableVerbose)
+                return;
+
             if (Cmdlet != null)
             {
                 Cmdlet.WriteVerbose(Message);
@@ -59,6 +81,27 @@ namespace PSFramework.Data
             {
                 runtime.AddCommand("Microsoft.PowerShell.Utility\\Write-Verbose").AddArgument(Message).Invoke();
             }
+        }
+
+        /// <summary>
+        /// Escapes a string's single quotes of all flavors.
+        /// </summary>
+        /// <param name="Text">The text to escape</param>
+        /// <returns>A properly escaped string</returns>
+        public string EscapeQuotes(string Text)
+        {
+            return CodeGeneration.EscapeSingleQuotedStringContent(Text);
+        }
+
+        /// <summary>
+        /// Escapes a string's single quotes of all flavors.
+        /// Ensures content is converted to text using PowerShell rules first.
+        /// </summary>
+        /// <param name="Value">The text to escape</param>
+        /// <returns>A properly escaped string</returns>
+        public string EscapeQuotes(object Value)
+        {
+            return CodeGeneration.EscapeSingleQuotedStringContent(LanguagePrimitives.ConvertTo<string>(Value));
         }
     }
 }